nodes/cryptonodes-status/requirements.txt
python-bitcoinlib 
jsonrpcclient[requests]
web3
